Extract withEnhancers tests to a separate file
